moving is_ipython_notebook to global utils
