+++ cartoframes/data/observatory/catalog/utils.py
@@ -1,20 +1,6 @@
 from .subscriptions import trigger_subscription
 from .subscription_info import fetch_subscription_info
-
-
-def is_ipython_notebook():
-    """
-    Detect whether we are in a Jupyter notebook.
-    """
-    try:
-        cfg = get_ipython().config
-        if 'IPKernelApp' in cfg:
-            return True
-        else:
-            return False
-    except NameError:
-        return False
-
+from ....utils.utils import is_ipython_notebook
 
 if is_ipython_notebook():
     from IPython.display import display
